// src/services/api/models/create-schedule-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Documentation
 * api documentation
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CreateScheduleDto
 */
export interface CreateScheduleDto {
  /**
   *
   * @type {number}
   * @memberof CreateScheduleDto
   */
  userId: number;
  /**
   *
   * @type {number}
   * @memberof CreateScheduleDto
   */
  doctorId: number;
  /**
   *
   * @type {string}
   * @memberof CreateScheduleDto
   */
  diagnosis?: string;
  /**
   *
   * @type {string}
   * @memberof CreateScheduleDto
   */
  comments?: string;
  /**
   *
   * @type {string}
   * @memberof CreateScheduleDto
   */
  startDate: string;
  /**
   *
   * @type {string}
   * @memberof CreateScheduleDto
   */
  endDate: string;
}