// src/services/api/models/doctor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Documentation
 * api documentation
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Doctor
 */
export interface Doctor {
  /**
   *
   * @type {string}
   * @memberof Doctor
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof Doctor
   */
  firstName: string;
  /**
   *
   * @type {string}
   * @memberof Doctor
   */
  middleName: string;
  /**
   *
   * @type {string}
   * @memberof Doctor
   */
  lastName: string;
}